requirements.txt
#Smart Vision Bot - Python Dependencies
# Core Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.1

# Machine Learning & AI
numpy==1.24.3
scikit-learn==1.3.0

# HTTP Requests & APIs
requests==2.31.0
urllib3==2.0.4

# Text-to-Speech
pyttsx3==2.90

# RSS Feed Parsing (for news fallback)
feedparser==6.0.10

# Environment Variables
python-dotenv==1.0.0

# Image Processing
Pillow==10.0.0

# Threading & Async
#threading-timer==0.1.0

# Data Handling
pickle-mixin==1.0.2

# System Utils
psutil==5.9.5

# Optional: For better performance
# Uncomment if you have issues with face_recognition
# cmake==3.27.4
# boost==0.1

# Development Dependencies (Optional)
# pytest==7.4.0
# flake8==6.0.0

# Platform-specific notes:
# Windows: All packages should install via pip
# macOS: May need Xcode command line tools for dlib
# Linux: May need: sudo apt-get install cmake libopenblas-dev liblapack-dev
# Raspberry Pi: Use pre-compiled wheels or allow extra time for compilation